fix: smb specific type removed from baseviewer class
