+++ src/core/sfu/protocolFactory.ts
@@ -1,11 +1,10 @@
 import { SfuType } from "./interface";
-import { SmbProtocol } from "./smb";
+import { SmbProtocol, SmbProtocolOptions } from "./smb";
 
 export default function (type: SfuType, sfuOptions) {
   switch (type) {
     case SfuType.smb:
-      return new SmbProtocol(sfuOptions);
-      break;
+      return new SmbProtocol(<SmbProtocolOptions>sfuOptions);
     default:
       throw new Error("Unsupported SFU type " + type);
   }
